requirements.txt
Django==4.2.9
Flask==2.3.2
gunicorn==21.2.0
huggingface-hub==0.20.1
ipykernel==6.28.0
ipython==8.18.1
langchain==0.0.353
langchain-community==0.0.7
langchain-core==0.1.4
langsmith==0.0.75
llama-index==0.9.41
openai==1.6.1
pypdf==4.0.1
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.29.0
streamlit-chat==0.0.1
Werkzeug==3.0.1
flask_session==0.6.0
flask_sqlalchemy==3.1.1
flask-swagger-ui==4.11.1